Move iterators into a separate header file
